code/scheduler/requirements.txt
django==5.0.6
django-rest-framework==0.1.0
ortools==9.10.4067